+++ pax-designer/src/glass/wireframe_editor/editor_generation/slot_control.rs
@@ -22,7 +22,7 @@ use crate::{
     model::{
         self,
         action::{
-            orm::{tree_movement::MoveNode, NodeLayoutSettings, SetNodeLayout},
+            orm::{movement::MoveNode, NodeLayoutSettings, SetNodeLayout},
             Action, ActionContext, RaycastMode,
         },
         input::InputEvent,
